requirements.txt
langchain 
langchain-community 
langchain-core
pypdf
tiktoken
openai
chromadb
# transformers
# torch
tf-keras
# tensorflow
sentence-transformers
# TMPDIR=/data/vincents/ pip install --cache-dir=/data/vincents/ sentence-transformers
